+++ pkg/R/lefko3-package.R
@@ -45,13 +45,13 @@
 #' @importFrom lme4 fixef glmer lmer ranef VarCorr
 #' @importFrom MASS glm.nb
 #' @importFrom methods is
-#' @importFrom MuMIn dredge r.squaredGLMM
+#' @importFrom MuMIn dredge
 #' @importFrom pscl zeroinfl
 #' @importFrom Rcpp evalCpp
 #' @importFrom SparseM as.matrix.csr image
 #' @importFrom stats getCall glm lm na.action na.fail na.omit rnorm sd setNames xtabs
 #' @importFrom stats as.formula median pchisq poisson var logLik
-#' @importFrom VGAM posnegbinomial pospoisson vglm
+#' @importFrom VGAM posnegbinomial pospoisson vglm predictvglm
 #' @useDynLib lefko3
 #' @name lefko3
 NULL